example module refactor contents: reducing example.rs into different mod.rs files in new example directory fine grain tests for all example game modules bugfix for general util general sum tree game ExampleGame trait to allow for module specific game construction trimmed comments and updated docstrings for utility module Visualizer trait to remove redundant visualize functions
