+++ src/game/mock/example/general_utility/mod.rs
@@ -1,0 +1,10 @@
+//! # Example Game Module - General Utility [WIP]
+//!
+//! This module contains games of type tree, acyclic, and cyclic that
+//! adhere to the following definitions:
+//!     General Utility - Player utilities are expressed numerically.
+
+/* SUBMODULES */
+
+pub mod general_sum;
+pub mod zero_sum;
