+++ src/game/mock/example/simple_utility/mod.rs
@@ -1,0 +1,10 @@
+//! # Example Game Module - Simple Utility
+//!
+//! This module contains games of type tree, acyclic, and cyclic that
+//! adhere to the following definitions:
+//!     Simple Utility - Player utilities are defined as WIN, LOSE, TIE, or DRAW.
+
+/* SUBMODULES */
+
+pub mod general_sum;
+pub mod zero_sum;
